+++ src/oracledb/impl/thin/statement.pyx
@@ -1,5 +1,5 @@
 #------------------------------------------------------------------------------
-# Copyright (c) 2020, 2024, Oracle and/or its affiliates.
+# Copyright (c) 2020, 2025, Oracle and/or its affiliates.
 #
 # This software is dual-licensed to you under the Universal Permissive License
 # (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl and Apache License
@@ -305,6 +305,7 @@ cdef class Statement:
         bint _no_prefetch
         bint _requires_define
         bint _return_to_cache
+        bint _is_nested
         bint _in_use
 
     cdef Statement copy(self):