+++ tests/test_6300_cursor_other_async.py
@@ -872,6 +872,47 @@ class TestCase(test_env.BaseAsyncTestCase):
         await self.cursor.execute("select * from TestJsonCols order by IntCol")
         self.assertEqual(await self.cursor.fetchall(), expected_data)
 
+    async def test_6352(self):
+        "6352 - test fetching nested cursors repeatedly"
+        sql = """
+            select
+                s.Description,
+                cursor(select 'Nested String for ' || s.Description from dual)
+            from
+                (
+                    select 'Top Level String 1' as Description
+                    from dual
+                    union all
+                    select 'Top Level String 2'
+                    from dual
+                    union all
+                    select 'Top Level String 3'
+                    from dual
+                    union all
+                    select 'Top Level String 4'
+                    from dual
+                    union all
+                    select 'Top Level String 5'
+                    from dual
+                ) s"""
+
+        for i in range(3):
+            with self.conn.cursor() as cursor:
+                cursor.arraysize = 10
+                await cursor.execute(sql)
+                desc, nested1 = await cursor.fetchone()
+                self.assertEqual(desc, "Top Level String 1")
+                nested_rows = await nested1.fetchall()
+                self.assertEqual(
+                    nested_rows, [("Nested String for Top Level String 1",)]
+                )
+                desc, nested2 = await cursor.fetchone()
+                self.assertEqual(desc, "Top Level String 2")
+                nested_rows = await nested2.fetchall()
+                self.assertEqual(
+                    nested_rows, [("Nested String for Top Level String 2",)]
+                )
+
 
 if __name__ == "__main__":
     test_env.run_test_cases()